Move @components/channel_list to @screens/home/channels_list/categories_list
