+++ app/screens/home/channel_list/categories_list/categories/body/channel_item/channel_item.test.tsx
@@ -8,12 +8,12 @@ import {MM_TABLES} from '@constants/database';
 import {renderWithIntlAndTheme} from '@test/intl-test-helper';
 import TestHelper from '@test/test_helper';
 
-import ChannelListItem from './channel_list_item';
+import ChannelItem from './channel_item';
 
 import type ChannelModel from '@typings/database/models/servers/channel';
 import type MyChannelModel from '@typings/database/models/servers/my_channel';
 
-describe('components/channel_list/categories/body/channel/item', () => {
+describe('components/channel_list/categories/body/channel_item', () => {
     let database: Database;
     let myChannel: MyChannelModel;
 
@@ -30,7 +30,7 @@ describe('components/channel_list/categories/body/channel/item', () => {
 
     it('should match snapshot', () => {
         const wrapper = renderWithIntlAndTheme(
-            <ChannelListItem
+            <ChannelItem
                 channel={{displayName: 'Hello!', type: 'G', shared: false, name: 'hello', deleteAt: 0} as ChannelModel}
                 isActive={false}
                 myChannel={myChannel}
