+++ app/screens/home/channel_list/categories_list/categories/body/channel_item/channel_item.tsx
