+++ app/screens/home/channel_list/categories_list/categories/body/index.ts
