+++ app/screens/home/channel_list/categories_list/categories/categories.test.tsx
@@ -7,7 +7,7 @@ import React from 'react';
 import {renderWithEverything} from '@test/intl-test-helper';
 import TestHelper from '@test/test_helper';
 
-import Categories from './';
+import Categories from '.';
 
 describe('components/channel_list/categories', () => {
     let database: Database;
