+++ app/screens/home/channel_list/categories_list/categories/error.tsx
@@ -5,10 +5,10 @@ import React, {useCallback, useState} from 'react';
 import {useIntl} from 'react-intl';
 
 import {retryInitialTeamAndChannel} from '@actions/remote/retry';
-import LoadingError from '@components/channel_list/loading_error';
+import LoadingError from '@components/loading_error';
 import {useServerDisplayName, useServerUrl} from '@context/server';
 
-const LoadTeamsError = () => {
+const LoadCategoriesError = () => {
     const {formatMessage} = useIntl();
     const serverUrl = useServerUrl();
     const serverName = useServerDisplayName();
@@ -26,11 +26,11 @@ const LoadTeamsError = () => {
     return (
         <LoadingError
             loading={loading}
-            message={formatMessage({id: 'load_teams_error.message', defaultMessage: 'There was a problem loading content for this server.'})}
+            message={formatMessage({id: 'load_categories_error.message', defaultMessage: 'There was a problem loading content for this server.'})}
             onRetry={onRetryTeams}
-            title={formatMessage({id: 'load_teams_error.title', defaultMessage: "Couldn't load {serverName}"}, {serverName})}
+            title={formatMessage({id: 'load_categories_error.title', defaultMessage: "Couldn't load categories for {serverName}"}, {serverName})}
         />
     );
 };
 
-export default LoadTeamsError;
+export default LoadCategoriesError;
