+++ app/screens/home/channel_list/channel_list.tsx
@@ -8,13 +8,13 @@ import {StyleSheet} from 'react-native';
 import Animated, {useAnimatedStyle, withTiming} from 'react-native-reanimated';
 import {Edge, SafeAreaView, useSafeAreaInsets} from 'react-native-safe-area-context';
 
-import ChannelList from '@components/channel_list';
 import FreezeScreen from '@components/freeze_screen';
 import TeamSidebar from '@components/team_sidebar';
 import {useTheme} from '@context/theme';
 import {useIsTablet} from '@hooks/device';
 import Channel from '@screens/channel';
 
+import CategoriesList from './categories_list';
 import Servers from './servers';
 
 type ChannelProps = {
@@ -84,7 +84,7 @@ const ChannelListScreen = (props: ChannelProps) => {
                         iconPad={canAddOtherServers}
                         teamsCount={props.teamsCount}
                     />
-                    <ChannelList
+                    <CategoriesList
                         iconPad={canAddOtherServers && props.teamsCount <= 1}
                         isTablet={isTablet}
                         teamsCount={props.teamsCount}
